fix: fix cf_dest tests
restructured tests and added node_modules as symlink
