+++ packages/ui5-middleware-cfdestination/test/boot.test.js
@@ -8,27 +8,37 @@ const request = require("supertest")
 const _request = require("axios")
 const server = require("@ui5/server").server
 const { spawn } = require("child_process")
+const copyUI5app = require("./_fs_app_util")
 
 const test = require("ava")
 
 const waitOn = require("wait-on")
 
-/**
- * copy showcase ui5 app for test purposes
- * and rm included ui5.yaml and xs-app.json
- *
- * @param {string} tmpDir path to copy ui5 app (that acts as the test app to)
- */
-async function copyUI5app(tmpDir) {
-	await fs.mkdir(tmpDir)
-	const filterFn = (src, _) => {
-		const yo = ["node_modules", "dist", "ui5.yaml", "xs-app.json"].find((node) => src.endsWith(node))
-		return yo === undefined ? true : false
-	}
-	await fs.copy(path.resolve(__dirname, "../../ui5-app"), tmpDir, {
-		filter: filterFn
-	})
-}
+// /**
+//  * copy showcase ui5 app for test purposes
+//  * and rm included ui5.yaml and xs-app.json
+//  *
+//  * @param {string} tmpDir path to copy ui5 app (that acts as the test app to)
+//  */
+// async function copyUI5app(tmpDir) {
+// 	await fs.mkdir(tmpDir)
+// 	const filterFn = (src, _) => {
+// 		const yo = ["node_modules", "dist", "ui5.yaml", "xs-app.json"].find((node) => src.endsWith(node))
+// 		if (yo === "node_modules") {
+// 			createSymlink(src, _)
+// 		}
+// 		return yo === undefined ? true : false
+// 	}
+// 	await fs.copy(path.resolve(__dirname, "../../ui5-app"), tmpDir, {
+// 		filter: filterFn
+// 	})
+// }
+
+// async function createSymlink(src, dest) {
+// 	await fs.ensureDir(path.dirname(dest))
+// 	await fs.remove(dest)
+// 	await fs.symlink(src, dest)
+// }
 
 test.beforeEach(async (t) => {
 	// copy ui5 app to a temp dir in test folder scope
