+++ packages/ui5-middleware-cfdestination/test/options.test.js
@@ -7,22 +7,7 @@ const { spawn } = require("child_process")
 const test = require("ava")
 const waitOn = require("wait-on")
 
-/**
- * copy showcase ui5 app for test purposes
- * and rm included ui5.yaml and xs-app.json
- *
- * @param {string} tmpDir path to copy ui5 app (that acts as the test app to)
- */
-async function copyUI5app(tmpDir) {
-	await fs.mkdir(tmpDir)
-	const filterFn = (src, _) => {
-		const yo = ["node_modules", "dist", "ui5.yaml", "xs-app.json"].find((node) => src.endsWith(node))
-		return yo === undefined ? true : false
-	}
-	await fs.copy(path.resolve(__dirname, "../../ui5-app"), tmpDir, {
-		filter: filterFn
-	})
-}
+const copyUI5app = require("./_fs_app_util")
 
 test.beforeEach(async (t) => {
 	// copy ui5 app to a temp dir in test folder scope
